verovista-app/requirements.txt
Flask>=2.3.0
Authlib>=1.2.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
bleach>=5.0.0
Flask-Talisman>=1.1.0
Flask-Limiter[memory]>=3.0.0
Werkzeug>=2.3.0
